+++ 3d_autofocus/minimize_entropy.h
@@ -2,6 +2,6 @@
 
 #include <cstdlib>
 
-void minimize_entropy(float *focusedImageR, float *focusedImageI, float
-        *minEntropy, float *origEntropy, float *Br, float *Bi, size_t K,
+void minimize_entropy(double *focusedImageR, double *focusedImageI, double
+        *minEntropy, double *origEntropy, double *Br, double *Bi, size_t K,
         size_t N);
